+++ app/routes/workspaces+/$id/campaigns/$selected_id/script/edit.route.tsx
@@ -293,7 +293,6 @@ export default function ScriptEditor() {
                     },
                   });
                 }}
-                surveys={[]}
               />
             </div>
           )}